feat-fix: fix update document when relateds is empty & add update for document file

+++ app/Service/Document/DocumentService.php
@@ -225,6 +225,71 @@ class DocumentService extends AppService implements AppServiceInterface
         }
     }
 
+    public function saveAndMark($documentInstance, $fileId) {
+        $fileOriginData = $this->fileTable->newQuery()->find($fileId);
+                
+        if($fileOriginData) {
+            $filePathOrigin = $fileOriginData->file_path;
+            try {
+                // Source file and watermark config 
+                $originalFile = 'storage/'.$filePathOrigin; 
+                $generateQR = \QrCode::format('png')->size(120)->merge('/public/images/square_ruang_mutu.png', .3)->errorCorrection('H')->generate(
+                    config('app.frontend_url') . '/view-file/doc/' . $documentInstance->id); 
+
+                $disk = env('UPLOAD_STORAGE', 'public');
+
+                $temporaryPath = 'uploads/temporary/';
+                $temporaryName = uniqid().'-'. time() . '.png';
+
+                $makeImage = \App\Service\OperationalStandard\OperationalStandardService::fromBase64(base64_encode($generateQR), $temporaryName);
+                $makeImage->storeAs($temporaryPath, $temporaryName, ['disk' => $disk]);
+                $text_image = 'storage/'.$temporaryPath.$temporaryName;
+                $pdf = new Fpdi(); 
+
+                if(file_exists(public_path($originalFile))){ 
+                    $pagecount = $pdf->setSourceFile($originalFile); 
+                }else{ 
+                    return $this->sendError(null,  'Terjadi kesalahan, silahkan ulangi beberapa saat');
+                } 
+    
+                for($i=1;$i<=$pagecount;$i++){ 
+                    $tpl = $pdf->importPage($i); 
+                    $size = $pdf->getTemplateSize($tpl); 
+                    $pdf->addPage(); 
+                    // $pdf->useTemplate($tpl, 1, 1, $size['width'], $size['height'], TRUE); 
+                    $pdf->useTemplate($tpl, null, 1, $size['width'], $size['height'], TRUE); 
+                    
+                    // $xxx_final = ($size['width']-35); 
+                    // $yyy_final = ($size['height']-275); 
+                    $xxx_final = ($size['width']-($size['width'] / 6)); 
+                    $yyy_final = ($size['height']-($size['height'] - 4)); 
+                    $pdf->Image($text_image, $xxx_final, $yyy_final, 0, 0, 'png'); 
+                } 
+                // Remove qrcode and temporary file From Storage
+                \Storage::disk($disk)->delete($temporaryPath.$temporaryName);
+                \Storage::disk($disk)->delete($filePathOrigin);
+                $fileOriginData->delete();
+
+                $newPdfName = FileUploadService::generateNewName() . '.pdf';
+                $makeNewPdf = \App\Service\OperationalStandard\OperationalStandardService::fromBase64(base64_encode($pdf->Output('S', $newPdfName)), $newPdfName);
+
+                $props = FileUploadService::directProcessFile($makeNewPdf);
+
+                $props['disk'] = $disk;
+                $props['group'] = 'document_files';
+                $props['fileable_id'] = $documentInstance->id;
+                $props['fileable_type'] = get_class($documentInstance);
+
+                $finalSave = FileUploadService::directSaveToDb($props);
+    
+            } catch (\Exception $e) {
+                // return $e->getMessage();
+                return $this->sendError(null,   $e->getMessage() ?? null);
+
+            }
+        }
+    }
+
     public function backendCreate($data)
     {
 
@@ -293,22 +358,31 @@ class DocumentService extends AppService implements AppServiceInterface
             
             $document->save();
 
-            $this->documentRelated->newQuery()->where('document_id', $id)->delete();
-            foreach($data['document_related'] as $doc) {
-                $this->documentRelated->newQuery()->create([
-                    'document_id'            =>  $document->id,
-                    'related_document_id'    =>  $doc,
-                ]);
+            if (isset($data['document_related']) && !empty($data['document_related'])) {
+                $this->documentRelated->newQuery()->where('document_id', $id)->delete();
+                foreach($data['document_related'] as $doc) {
+                    $this->documentRelated->newQuery()->create([
+                        'document_id'            =>  $document->id,
+                        'related_document_id'    =>  $doc,
+                    ]);
+                }
             }
             
-            $this->programRelated->newQuery()->where('document_id', $id)->delete();
             if (isset($data['program_related'])) {
+                $this->programRelated->newQuery()->where('document_id', $id)->delete();
                 foreach($data['program_related'] as $program) {
                     $this->programRelated->newQuery()->create([
                         'document_id'            =>  $document->id,
                         'program_id'             =>  $program,
                     ]);
                 }
+            }
+
+            if (isset($data['file_id']) && !empty($data['file_id'])) {
+                if($document->file) {
+                    $document->file->delete();
+                }
+                $this->saveAndMark($document, $data['file_id']);
             }
 
             \DB::commit(); // commit the changes
